requirements.txt
Django=4.2.4
mysql-connector-python=8.1.0
mysqlclient=2.2.0